+++ spykeutils/__init__.py
@@ -73,7 +73,7 @@
     :members:
 """
 
-__version__ = '0.2.1'
+__version__ = '0.2.2'
 
 
 class SpykeException(Exception):