+++ CHANGELOG.rst
@@ -1,6 +1,3 @@
-Changelog
-=========
-
 Version 0.2.1
 -------------
 * Added "Home" and "Pan" tools for plots (useful when no middle mouse button
@@ -10,6 +7,11 @@ Version 0.2.1
   changed.
 * New parameter for DataProvider AnalogSignal methods: AnalogSignalArrays can
   be automatically converted and returned.
+* Significantly improved speed of spike density estimation and optimal kernel
+  size calculation.
+* Spike sorting quality assessment using gaussian clusters is now possible
+  without prewhitening spikes or providing prewhitened means.
+* Renamed "spyke-plugin" script to "spykeplugin"
 
 Version 0.2.0
 -------------